+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/impl/SnapshotTransaction.java
@@ -120,13 +120,10 @@ import com.palantir.lock.AtlasCellLockDescriptor;
 import com.palantir.lock.AtlasRowLockDescriptor;
 import com.palantir.lock.LockDescriptor;
 import com.palantir.lock.v2.ClientLockingOptions;
-import com.palantir.lock.v2.ImmutableLockRequest;
 import com.palantir.lock.v2.LockRequest;
 import com.palantir.lock.v2.LockResponse;
 import com.palantir.lock.v2.LockToken;
 import com.palantir.lock.v2.TimelockService;
-import com.palantir.lock.v2.WaitForLocksRequest;
-import com.palantir.lock.v2.WaitForLocksResponse;
 import com.palantir.logsafe.Preconditions;
 import com.palantir.logsafe.SafeArg;
 import com.palantir.logsafe.UnsafeArg;
@@ -204,6 +201,18 @@ public class SnapshotTransaction extends AbstractTransaction
     @VisibleForTesting
     static final int MIN_BATCH_SIZE_FOR_DISTRIBUTED_LOAD = 100;
 
+    protected ListenableFuture<Map<Long, Long>> getCommitTimestamps(
+            TableReference tableRef,
+            Iterable<Long> startTimestamps,
+            boolean shouldWaitForCommitterToComplete,
+            AsyncTransactionService asyncTransactionService) {
+        return commitTimestampLoader.getCommitTimestamps(
+                tableRef,
+                startTimestamps,
+                shouldWaitForCommitterToComplete,
+                asyncTransactionService);
+    }
+
     private enum State {
         UNCOMMITTED,
         COMMITTED,
@@ -260,9 +269,8 @@ public class SnapshotTransaction extends AbstractTransaction
     protected final Supplier<TransactionConfig> transactionConfig;
     protected final TableLevelMetricsController tableLevelMetricsController;
     protected final SuccessCallbackManager successCallbackManager = new SuccessCallbackManager();
-    private final Supplier<Long> lastSeenCommitTs;
     protected final TransactionSchemaManager transactionSchemaManager;
-
+    private final CommitTimestampLoader commitTimestampLoader;
     protected volatile boolean hasReads;
 
     /**
@@ -329,7 +337,15 @@ public class SnapshotTransaction extends AbstractTransaction
         this.validateLocksOnReads = validateLocksOnReads;
         this.transactionConfig = transactionConfig;
         this.tableLevelMetricsController = tableLevelMetricsController;
-        this.lastSeenCommitTs = lastSeenCommitTs;
+        this.commitTimestampLoader = new CommitTimestampLoader(timestampValidationReadCache,
+                immutableTimestampLock,
+                transactionSchemaManager,
+                this::getStartTimestamp,
+                transactionConfig,
+                metricsManager,
+                timelockService,
+                immutableTimestamp,
+                lastSeenCommitTs);
     }
 
     protected TransactionScopedCache getCache() {
@@ -1529,7 +1545,8 @@ public class SnapshotTransaction extends AbstractTransaction
         Set<Long> valuesStartTimestamps = getStartTimestampsForValues(rawResults.values());
 
         return Futures.transformAsync(
-                getCommitTimestamps(tableRef, valuesStartTimestamps, true, asyncTransactionService),
+                commitTimestampLoader.getCommitTimestamps(tableRef, valuesStartTimestamps, true,
+                        asyncTransactionService),
                 commitTimestamps -> collectCellsToPostFilter(
                         tableRef,
                         rawResults,
@@ -2017,21 +2034,6 @@ public class SnapshotTransaction extends AbstractTransaction
         throwIfPreCommitConditionInvalid(timestamp);
     }
 
-    private void throwIfTransactionsTableSweptBeyondReadOnlyTxn() {
-        long startTs = getStartTimestamp();
-        // The schema version of current transaction does not matter. If the current transaction does not hold
-        // immutableTs lock, and we were previously on schema 4 for a range of transactions, we cannot know the state
-        // of those writes consistently if sweep has progressed.
-
-        if (immutableTimestampLock.isEmpty()) {
-            Preconditions.checkState(
-                    lastSeenCommitTs.get() < startTs,
-                    "Transactions table has been swept beyond current start timestamp, therefore, we cannot"
-                            + " consistently values accessible to this transactions. This can happen if the transaction"
-                            + " has been alive for more than an hour and is expected to be transient.");
-        }
-    }
-
     private boolean isReadingSweepableTransaction(long startTs) {
         return transactionSchemaManager.getTransactionsSchemaVersion(startTs)
                 == TransactionConstants.TTS_TRANSACTIONS_SCHEMA_VERSION;
@@ -2339,10 +2341,7 @@ public class SnapshotTransaction extends AbstractTransaction
 
         // TODO(fdesouza): Revert this once PDS-95791 is resolved.
         long lockAcquireTimeoutMillis = currentTransactionConfig.getLockAcquireTimeoutMillis();
-        LockRequest request = ImmutableLockRequest.of(
-                lockDescriptors,
-                lockAcquireTimeoutMillis,
-                Optional.ofNullable(getStartTimestampAsClientDescription(currentTransactionConfig)));
+        LockRequest request = LockRequest.of(lockDescriptors, lockAcquireTimeoutMillis);
 
         RuntimeException stackTraceSnapshot = new SafeRuntimeException("I exist to show you the stack trace");
         LockResponse lockResponse = timelockService.lock(
@@ -2405,61 +2404,6 @@ public class SnapshotTransaction extends AbstractTransaction
         return result;
     }
 
-    /**
-     * We will block here until the passed transactions have released their lock.  This means that the committing
-     * transaction is either complete or it has failed and we are allowed to roll it back.
-     */
-    private void waitForCommitToComplete(Iterable<Long> startTimestamps) {
-        Set<LockDescriptor> lockDescriptors = new HashSet<>();
-        for (long start : startTimestamps) {
-            if (start < immutableTimestamp) {
-                // We don't need to block in this case because this transaction is already complete
-                continue;
-            }
-            lockDescriptors.add(AtlasRowLockDescriptor.of(
-                    TransactionConstants.TRANSACTION_TABLE.getQualifiedName(),
-                    TransactionConstants.getValueForTimestamp(start)));
-        }
-
-        if (lockDescriptors.isEmpty()) {
-            return;
-        }
-
-        waitFor(lockDescriptors);
-    }
-
-    private void waitFor(Set<LockDescriptor> lockDescriptors) {
-        TransactionConfig currentTransactionConfig = transactionConfig.get();
-        String startTimestampAsDescription = getStartTimestampAsClientDescription(currentTransactionConfig);
-
-        // TODO(fdesouza): Revert this once PDS-95791 is resolved.
-        long lockAcquireTimeoutMillis = currentTransactionConfig.getLockAcquireTimeoutMillis();
-        WaitForLocksRequest request =
-                WaitForLocksRequest.of(lockDescriptors, lockAcquireTimeoutMillis, startTimestampAsDescription);
-        WaitForLocksResponse response = timelockService.waitForLocks(request);
-        if (!response.wasSuccessful()) {
-            log.error(
-                    "Timed out waiting for commits to complete. Timeout was {} ms. First ten locks were {}.",
-                    SafeArg.of("requestId", request.getRequestId()),
-                    SafeArg.of("acquireTimeoutMs", lockAcquireTimeoutMillis),
-                    SafeArg.of("numberOfDescriptors", lockDescriptors.size()),
-                    UnsafeArg.of("firstTenLockDescriptors", Iterables.limit(lockDescriptors, 10)));
-            throw new TransactionLockAcquisitionTimeoutException("Timed out waiting for commits to complete.");
-        }
-    }
-
-    /**
-     * TODO(fdesouza): Remove this once PDS-95791 is resolved.
-     *
-     * @deprecated Remove this once PDS-95791 is resolved.
-     */
-    @Deprecated
-    @Nullable
-    private String getStartTimestampAsClientDescription(TransactionConfig currentTransactionConfig) {
-        return currentTransactionConfig.attachStartTimestampToLockRequestDescriptions()
-                ? Long.toString(getStartTimestamp())
-                : null;
-    }
 
     ///////////////////////////////////////////////////////////////////////////
     /// Commit timestamp management
@@ -2475,124 +2419,8 @@ public class SnapshotTransaction extends AbstractTransaction
 
     private Map<Long, Long> getCommitTimestampsSync(
             @Nullable TableReference tableRef, Iterable<Long> startTimestamps, boolean waitForCommitterToComplete) {
-        return AtlasFutures.getUnchecked(getCommitTimestamps(
+        return AtlasFutures.getUnchecked(commitTimestampLoader.getCommitTimestamps(
                 tableRef, startTimestamps, waitForCommitterToComplete, immediateTransactionService));
-    }
-
-    /**
-     * Returns a map from start timestamp to commit timestamp.  If a start timestamp wasn't committed, then it will be
-     * missing from the map.  This method will block until the transactions for these start timestamps are complete.
-     */
-    protected ListenableFuture<Map<Long, Long>> getCommitTimestamps(
-            @Nullable TableReference tableRef,
-            Iterable<Long> startTimestamps,
-            boolean shouldWaitForCommitterToComplete,
-            AsyncTransactionService asyncTransactionService) {
-        if (Iterables.isEmpty(startTimestamps)) {
-            return Futures.immediateFuture(ImmutableMap.of());
-        }
-
-        boolean isReadOnly = immutableTimestampLock.isEmpty();
-
-        Set<Long> pendingGets = new HashSet<>();
-        Set<Long> sweepableGets = new HashSet<>();
-        Map<Long, Long> startToCommitTimestamps = new HashMap<>();
-
-        for (Long startTs : startTimestamps) {
-            if (isReadOnly && isReadingSweepableTransaction(startTs)) {
-                sweepableGets.add(startTs);
-                pendingGets.add(startTs);
-            } else {
-                Long commitTs = timestampValidationReadCache.getCommitTimestampIfPresent(startTs);
-                if (commitTs == null) {
-                    pendingGets.add(startTs);
-                } else {
-                    startToCommitTimestamps.put(startTs, commitTs);
-                }
-            }
-        }
-
-        if (pendingGets.isEmpty()) {
-            return Futures.immediateFuture(startToCommitTimestamps);
-        }
-
-        // Before we do the reads, we need to make sure the committer is done writing.
-        if (shouldWaitForCommitterToComplete) {
-            waitForCommitterToComplete(tableRef, startTimestamps);
-        }
-
-        traceGetCommitTimestamps(tableRef, pendingGets);
-
-        if (pendingGets.size() > transactionConfig.get().getThresholdForLoggingLargeNumberOfTransactionLookups()) {
-            logLargeNumberOfTransactions(tableRef, pendingGets);
-        }
-
-        return Futures.transform(
-                loadCommitTimestamps(asyncTransactionService, pendingGets),
-                rawResults -> {
-                    for (Map.Entry<Long, Long> e : rawResults.entrySet()) {
-                        if (e.getValue() != null) {
-                            Long startTs = e.getKey();
-                            Long commitTs = e.getValue();
-                            startToCommitTimestamps.put(startTs, commitTs);
-                            timestampValidationReadCache.putAlreadyCommittedTransaction(startTs, commitTs);
-                        }
-                    }
-                    if (!sweepableGets.isEmpty()) {
-                        throwIfTransactionsTableSweptBeyondReadOnlyTxn();
-                    }
-                    return startToCommitTimestamps;
-                },
-                MoreExecutors.directExecutor());
-    }
-
-    private void waitForCommitterToComplete(@Nullable TableReference tableRef, Iterable<Long> startTimestamps) {
-        Timer.Context timer = getTimer("waitForCommitTsMillis").time();
-        waitForCommitToComplete(startTimestamps);
-        long waitForCommitTsMillis = TimeUnit.NANOSECONDS.toMillis(timer.stop());
-
-        if (tableRef != null) {
-            perfLogger.debug(
-                    "Waited to get commit timestamps when reading from a known table.",
-                    SafeArg.of("commitTsMillis", waitForCommitTsMillis),
-                    LoggingArgs.tableRef(tableRef));
-        } else {
-            perfLogger.debug("Waited to get commit timestamps.", SafeArg.of("commitTsMillis", waitForCommitTsMillis));
-        }
-    }
-
-    private void traceGetCommitTimestamps(@Nullable TableReference tableRef, Set<Long> gets) {
-        if (tableRef != null) {
-            log.trace(
-                    "Getting commit timestamps for a read while reading table.",
-                    SafeArg.of("numTimestamps", gets.size()),
-                    LoggingArgs.tableRef(tableRef));
-            return;
-        }
-
-        log.trace("Getting commit timestamps.", SafeArg.of("numTimestamps", gets.size()));
-    }
-
-    private void logLargeNumberOfTransactions(@Nullable TableReference tableRef, Set<Long> gets) {
-        log.info(
-                "Looking up a large number of transactions.",
-                SafeArg.of("numberOfTransactionIds", gets.size()),
-                tableRef == null ? SafeArg.of("tableRef", "no_table") : LoggingArgs.tableRef(tableRef));
-    }
-
-    private static ListenableFuture<Map<Long, Long>> loadCommitTimestamps(
-            AsyncTransactionService asyncTransactionService, Set<Long> startTimestamps) {
-        // distinguish between a single timestamp and a batch, for more granular metrics
-        if (startTimestamps.size() == 1) {
-            Long singleTs = startTimestamps.iterator().next();
-            return Futures.transform(
-                    asyncTransactionService.getAsync(singleTs),
-                    commitTsOrNull ->
-                            commitTsOrNull == null ? ImmutableMap.of() : ImmutableMap.of(singleTs, commitTsOrNull),
-                    MoreExecutors.directExecutor());
-        } else {
-            return asyncTransactionService.getAsync(startTimestamps);
-        }
     }
 
     /**
